+++ config/index.js
@@ -5,6 +5,14 @@
 require('dotenv').config();
 // const nodeEnv = process.env.NODE_ENV || 'development';
 
+// Check requirement for minimum node version
+const minNodeVersion = 18;
+if (parseInt(process.version.replace('v', '').split('.')[0]) < minNodeVersion) {
+  console.error('Error: this program requires node version ' +
+    minNodeVersion.toString() + ' or greater.');
+  process.exit(1);
+}
+
 exports.app = {
   pidFilename: process.env.APP_PID_FILENAME || '',
   collectIntervalSeconds: parseInt(process.env.APP_COLLECT_INTERVAL_SECONDS || '60'),